requirements.txt
docling

nest_asyncio

# Milvus
pymilvus
milvus-lite

## Weaviate DB
# weaviate-client

litellm

## --- llama-index
llama-index
llama-index-embeddings-huggingface
llama-index-llms-litellm
llama-index-vector-stores-milvus

## --- UI - flask
flask

## utils
humanfriendly
pandas
mimetypes-magic
tqdm
ipykernel


# UI - chainlit
chainlit

# Environment variables
python-dotenv